refactor: Логика воркера разбита по нескольким файлам
